requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dotenv==0.9.9
flasgger==0.9.7.1
Flask==3.1.0
Flask-HTTPAuth==4.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.2
numpy==2.2.3
openai==1.64.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
